Fixed reports batch formatter
