+++ reports_batch_formatter/formatter.py
@@ -1,13 +1,16 @@
+import argparse
 import tarfile
 import tempfile
+import sys
+
 from os import path, listdir, makedirs, unlink
 from os.path import join
 
 import logging
 from shutil import copy
 
+sys.path.insert(0, '../')
 from processor import reports
-from reports_batch_formatter import parse_args
 
 logger = logging.getLogger(__name__)
 temp_dir = None
@@ -37,6 +40,18 @@ def reshape_results(working_directory):
         for report in reports_handler.processable_reports:
             unlink(report.file_path)
 
+def parse_args(raw_args=None):
+    parser = argparse.ArgumentParser(description='Script to shape the report files from the tix-time-condenser into '
+                                                 'the batches. This is to imitate the way the tix-time-processor takes '
+                                                 'the files and computes them by separating them into different '
+                                                 'directories. The idea behind this is to use the files for '
+                                                 'exploratory analysis.')
+    parser.add_argument('--source_directory',
+                        help='The path to the directory where the reports are.')
+    parser.add_argument('--output', '-o', action='store', default='batch-test-report.tar.gz', type=str,
+                        help='The name of the output file. By default "batch-test-report.tar.gz".')
+    args = parser.parse_args(raw_args)
+    return args
 
 if __name__ == "__main__":
     args = parse_args()
